Split entity trait
